pyproject.toml
[tool.poetry]
name = "inaturalist-uploads"
version = "0.1.0"
description = ""
authors = ["Evan Hamrick"]

[tool.poetry.dependencies]
python = "^3.7"
pyinaturalist = "^0.8.0"
Pillow = "^6.1"
lxml = "^4.4"
gpxpy = "^1.3"
pytz = "^2019.2"
PySimpleGUI = "^4.1"
keyring = "^19.0"
pendulum = "^2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
